Reset the ‘reset forgotten password’ form to its original state …and again retrieve the user email based on the supplied token.
Laravel5.3 stored the password reset token in the database and then sent
it to the user in an email, This made it possible to query the password
resets table directly to retrieve the user email using the token.

Laravel5.4 is using hashing and ‘password_verify’ to confirm the match
between the token in the database and the token from the email. This
commit reads through the password_reset table, checking each row for a
password match, and returns the associated email when it finds one.